GVT-2627: Add PreAuthorizations to LayoutDesignController endpoints GVT-2627: Add front-end privilege checks to workspace-related UI components GVT-2627: Separate workspace-related front-end stuff into its own component (WorkspaceSelection)
